feat: Add ElectionView component and update AppRouter
The ElectionView component is added to handle the election view in the application. The AppRouter is updated to include the new route for the ElectionView component.

chore: Update dependencies and add font styles for main menu

The dependencies are updated to their latest versions. Font styles are added for the main menu to improve the visual appearance.

fix: Update subtitle in vista_main

The subtitle in the vista_main component is updated to reflect the correct era of post-quantum cryptography.

chore: Update npm dependency to latest stable version

The npm dependency is updated to the latest stable version.

chore: Update landing page banner color per client request

The landing page banner color is updated based on the client's request.

fix: Increase size and position of MaterialUISwitch in CifradoView

The size and position of the MaterialUISwitch component in the CifradoView are increased and adjusted for better visibility and user experience.

feat: improve page load with lazy loading for images

Lazy loading for images is implemented to improve the page load performance.

Fix bug preventing submitting the signup form

A bug that was preventing the signup form from being submitted is fixed.